Refactor server, create API for interfacing with our DB
