+++ index.js
@@ -7,6 +7,7 @@ const io = require('socket.io')(http);
 const port = process.env.PORT || 30000;
 const admin = require('firebase-admin');
 const serviceAccount = require('./key.json');
+const dbAPI = require('./dbAPI');
 
 http.listen( port, function () {
     console.log('listening on port', port);
@@ -51,92 +52,14 @@ io.on('connection', function(socket){
     });
 
     // User log in: Check if new or returning user
-    socket.on('login', function(user) {
-
-        // Query DB for user by email
-        const ref = admin.firestore().collection('users').doc(user.email);
-        ref.get().then(function(doc) {
-
-            // User is not yet registered
-            if(!doc.exists) {
-
-                // Register user
-                admin.firestore().collection('users').doc(user.email).set({
-                    email: user.email,      // Email = Key
-                    name: user.displayName, // Username
-                    singleHighScore: 0,     // Single player high score
-                    multiHighScore: 0,      // Multi player high score
-                    chatColor: '#ffffff',   // Chat message color
-                    imageLink: null,        // User profile image URL
-                    gameInProgress: false,  // Is the user in a game right now?
-                    savedGame: null         // Previously saved single player game
-                })
-                    // Success: redirect to home page
-                    .then(function() { console.log('Success: ' + user.email + ' registered.'); })
-
-                    // Error
-                    .catch(function(error) { console.error('Error: ', error); });
-            }
-
-            // Login done, redirect to home page
-            socket.emit('login', '/home.html');
-        });
-    });
+    socket.on('login', function(user) { dbAPI.registerUser(admin, socket, user); });
 
     // User data requested
-    socket.on('get user', function(user) {
-
-        // Get reference to user in DB and fetch data
-        const ref = admin.firestore().collection('users').doc(user.email);
-        ref.get()
-
-            // Success: Send user data
-            .then(function(doc) { socket.emit('get user', doc.data()); })
-
-            // Error
-            .catch(function(error) { console.error("Error: ", error); });
-    });
+    socket.on('get user', function(user) { dbAPI.getUser(admin, socket, user); });
 
     // User name change requested
-    socket.on('new name', function(user, newName) {
-
-        // Get reference to user in DB and update it
-        const ref = admin.firestore().collection('users').doc(user.email);
-        let data = {
-            email: user.email,                      // Email = Key
-            name: newName,                          // New username
-            singleHighScore: user.singleHighScore,  // Single player high score
-            multiHighScore: user.multiHighScore,    // Multi player high score
-            chatColor: user.chatColor,              // Chat message color
-            imageLink: user.imageLink,              // User profile image URL
-            gameInProgress: user.gameInProgress,    // Is the user in a game right now?
-            savedGame: user.savedGame               // Previously saved single player game
-        };
-        ref.set(data)
-
-            // Success
-            .then(function () { socket.emit('new name', newName); })
-
-            // Error
-            .catch(function(error) { console.error('Error: ', error); });
-    });
+    socket.on('new name', function(user, newName) { dbAPI.changeUserName(admin, user, newName, socket); });
 
     // Leaderboard data requested
-    socket.on('get leaderboard', function() {
-
-        // Get top 10 user scores from DB
-        admin.firestore().collection('users')
-            .orderBy('singleHighScore', 'desc')
-            .limit(10)
-            .get()
-
-            // Success: Send each leaderboard entry to the client
-            // TODO send the whole thing instead of each entry individually
-            .then(function(querySnapshot) {
-                querySnapshot.forEach(function(doc) { socket.emit('get leaderboard', doc.data()); });
-            })
-
-            // Error
-            .catch(function(error) { console.error("Error: ", error) });
-    })
+    socket.on('get leaderboard', function() { dbAPI.getLeaderboard(admin, socket); });
 });
