+++ public/home.html
@@ -7,7 +7,7 @@ DOCUMENT METADATA:                                                              
 <head>
     <meta charset="UTF-8">
     <title>Online Last Letter - Users</title>
-    <link rel="stylesheet" type="text/css" href="../css/style.css">
+    <link rel="stylesheet" type="text/css" href="style.css">
 
 <!--====================================================================================================================
 SCRIPTS -->
@@ -29,8 +29,8 @@ SCRIPTS -->
     <script src="/socket.io/socket.io.js"></script>
 
     <!--Components-->
-    <script src="../js/client.js"></script>
-    <script src="../js/user-profile.js"></script>
+    <script src="client.js"></script>
+    <script src="user-profile.js"></script>
 </head>
 
 <body>
