+++ public/user-auth.html
@@ -6,7 +6,7 @@ DOCUMENT METADATA:                                                              
 
 <head>
     <title>LAST LETTER</title>                                      <!-- Define the title in the browser -->
-    <link rel="stylesheet" href="../css/style.css">                        <!-- Link to an external style file -->
+    <link rel="stylesheet" href="style.css">                        <!-- Link to an external style file -->
     <meta name="viewport" content ="width=device-width, initial-scale=1">
 
     <!--====================================================================================================================
