+++ public/user-auth.js
